requirements.txt
Django==5.2
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-environ
PyJWT
Pillow
pymysql